Refactor: update styles name for for better standardization
