+++ mobile/styles/IndexStyles.ts
