+++ mobile/styles/auth/PaymentStyles.ts
