+++ mobile/styles/auth/RegisterStyles.ts
