+++ mobile/styles/stock/ProductListStyles.ts
