+++ mobile/styles/stock/StockStyles.ts
